+++ packages/@webex/plugin-meetings/test/unit/spec/meeting-info/index.js
@@ -26,7 +26,7 @@ describe('plugin-meetings', () => {
 
     describe('#fetchMeetingInfo', () => {
       const checkResolvedFetchMeetingInfo = async ({meetingId, sendCAevents, shouldSendCAMetrics}) => {
-        const body = {meetingKey: '1234323', url: 'url-123'};
+        const body = {meetingKey: '1234323', url: 'url-123', confID: '123', meetingId: '321'};
 
         sinon
           .stub(MeetingInfoUtil, 'generateOptions')
@@ -66,6 +66,8 @@ describe('plugin-meetings', () => {
             },
             options: {
               meetingId,
+              webexConferenceIdStr: '123',
+              globalMeetingId: '321'
             },
           });
         } else {